rename pattern.h
