+++ source/core/plugin/CMakeLists.txt
@@ -8,7 +8,7 @@ set(SOURCES
 
 set(HEADERS
   ${dir}/container.h
-  ${dir}/pattern.h
+  ${dir}/Pattern.h
   ${dir}/Plugin.h
   ${dir}/precise_float.h
   ${dir}/Setting.h
