+++ source/core/plugin/Pattern.h
